==> time-series-sales-prediction/Step4.1.Models-Linear-Models.ipynb <==
# Sales Prediction for Time Series Data

## Part 4: Models (Linear Models)

For the linear models, the first step is to see which scaler performs the best.  After selecting the scaler, comparisons will be made between the linear regression with no regularization, ridge regression, and elastic net.

Submissions are evaluated by root mean squared error (RMSE). True target values are clipped into [0,20] range.

import numpy as np
import pandas as pd 
from sklearn import linear_model
from sklearn.metrics import mean_squared_error
from sklearn.model_selection import GridSearchCV
from sklearn.externals import joblib
import lightgbm as lgb
import os
import time 
import datetime
import seaborn as sns
import matplotlib.pyplot as plt
import random
import gc

%matplotlib inline 

y_train = joblib.load('../data/y_train_lev_1.pkl')
y_test = joblib.load('../data/y_val_lev_1.pkl')
y_train_full = joblib.load('../data/y_train_full.pkl')

def submission(model, X_test):
    '''
    make submission file
    arguments:    model = model name 
                  X_test= X_test name
    return: a file saved in directory with timestamp
    '''
    # model prediction
    pred = model.predict(X_test)
    print('mean before clipping: ', pred.mean())
    pred = pred.clip(0,20)
    print('mean after clipping: ', pred.mean())

    # create prediction dataframe
    ID = joblib.load('ID.pkl')
    predDF = pd.DataFrame() 
    predDF['ID'] = ID
    predDF['item_cnt_month'] = pred
    print(predDF.head())

    # write dataframe to csv
    ts = time.time()
    st = datetime.datetime.fromtimestamp(ts).strftime('%m%d_%H.%M')
    print('submission_' + st + '.csv')
    
    predDF.to_csv(header=True, index=False, path_or_buf = 'submission_' + st + '.csv')
    
    return None

def linearReg(X_train, y_train, X_test, y_test):
    X_train = X_train.drop(['shop_id','item_id', 'item_category_id'], axis = 1)
    X_test = X_test.drop(['shop_id','item_id', 'item_category_id'], axis = 1)
    model = linear_model.LinearRegression()
    model.fit(X_train.values, y_train)
    pred = model.predict(X_test.values)
    pred = pred.clip(0, 20)

    MSETest = mean_squared_error(y_test, pred)
    RMSETest = np.sqrt(MSETest)
    
    MSETrain = mean_squared_error(y_train, model.predict(X_train.values).clip(0, 20))
    RMSETrain = np.sqrt(MSETrain)
    
    print('Train RMSE:', RMSETrain)
    print('Test RMSE:', RMSETest)
    print('')
    print('y_test mean: ', y_test.mean())
    print('prediction mean: ', pred.mean())
    
    return model

### Compare different scalers

Different scaler are used to train the linear model.    

#### No Scaling

X_train = pd.read_pickle('../data/X_train_lev_1_noScaler')
X_test = pd.read_pickle('../data/X_test_lev_1_noScaler')

model = linearReg(X_train, y_train, X_test, y_test)

st = datetime.datetime.fromtimestamp(time.time()).strftime('%m%d_%H.%M')
joblib.dump(model, 'lr_no_scale_' + st + '.pkl')
print('lr_no_scale_' + st + '.pkl')

#### Standard Scaler

X_train = pd.read_pickle('../data/X_train_lev_1_standardScaler')
X_test = pd.read_pickle('../data/X_test_lev_1_standardScaler')

model = linearReg(X_train, y_train, X_test, y_test)

st = datetime.datetime.fromtimestamp(time.time()).strftime('%m%d_%H.%M')
joblib.dump(model, 'lr_standard_scale_' + st + '.pkl')
print('lr_standard_scale_' + st + '.pkl')

#### RobustScaler

X_train = pd.read_pickle('../data/X_train_lev_1_RobustScaler')
X_test = pd.read_pickle('../data/X_test_lev_1_RobustScaler')

model = linearReg(X_train, y_train, X_test, y_test)

st = datetime.datetime.fromtimestamp(time.time()).strftime('%m%d_%H.%M')
joblib.dump(model, 'lr_robust_scale_' + st + '.pkl')
print('lr_robust_scale_' + st + '.pkl')

#### MinMax Scaler

X_train = pd.read_pickle('../data/X_train_lev_1_MinMaxScaler')
X_test = pd.read_pickle('../data/X_test_lev_1_MinMaxScaler')

model = linearReg(X_train, y_train, X_test, y_test)

st = datetime.datetime.fromtimestamp(time.time()).strftime('%m%d_%H.%M')
joblib.dump(model, 'lr_minmax_scale_' + st + '.pkl')
print('lr_minmax_scale_' + st + '.pkl')

#### Quantile Scaler

X_train = pd.read_pickle('../data/X_train_lev_1_QuantileScaler')
X_test = pd.read_pickle('../data/X_test_lev_1_QuantileScaler')

model = linearReg(X_train, y_train, X_test, y_test)

st = datetime.datetime.fromtimestamp(time.time()).strftime('%m%d_%H.%M')
joblib.dump(model, 'lr_quantile_scale_' + st + '.pkl')
print('lr_quantile_scale_' + st + '.pkl')

### Compare with Ridge Regression and Elastic Net

#### Ridge Regression - L2 regularization
Alpha of 0 corresponds to linear regression.  

X_train = pd.read_pickle('../data/X_train_lev_1_MinMaxScaler')
X_test = pd.read_pickle('../data/X_test_lev_1_MinMaxScaler')
X_train = X_train.drop(['shop_id','item_id', 'item_category_id'], axis = 1)
X_test = X_test.drop(['shop_id','item_id', 'item_category_id'], axis = 1)

ridgeResult = {}
for alpha in [0, 1e-10, 1e-5, 1e-3, 1e-2, 0.5, 1, 10]:    
    # define model
    model = linear_model.Ridge(alpha = alpha, normalize = True)
    model.fit(X_train.values, y_train)
    
    #make prediction
    pred = model.predict(X_test.values)
    pred = pred.clip(0, 20)

    # calculate score
    MSETest = mean_squared_error(y_test, pred)
    RMSETest = np.sqrt(MSETest)
    MSETrain = mean_squared_error(y_train, model.predict(X_train.values))
    RMSETrain = np.sqrt(MSETrain)
    
    # save results
    ridgeResult[alpha] = RMSETest
    
    st = datetime.datetime.fromtimestamp(time.time()).strftime('%m%d_%H.%M')
    joblib.dump(model, 'ridge_' + str(alpha) + '_' + st + '.pkl')
    
    print('ridge_' + str(alpha) + '_' + st + '.pkl')
    print('Alpha:', alpha, 'Train RMSE:', RMSETrain, 'Test RMSE:', RMSETest)
    

print('The alpha with best RMSE: ', min(ridgeResult, key=ridgeResult.get), 'RMSE:', ridgeResult[min(ridgeResult, key=ridgeResult.get)])

#### Elastic Net

l1_ratio = 0 the penalty is an L2 penalty. 
<br> For l1_ratio = 1 it is an L1 penalty. 
<br>For 0 < l1_ratio < 1, the penalty is a combination of L1 and L2.

X_train = pd.read_pickle('../data/X_train_lev_1_MinMaxScaler')
X_test = pd.read_pickle('../data/X_test_lev_1_MinMaxScaler')
X_train = X_train.drop(['shop_id','item_id', 'item_category_id'], axis = 1)
X_test = X_test.drop(['shop_id','item_id', 'item_category_id'], axis = 1)

for l1_ratio in [0.01, 0.05, .1, .5, .7, .9, 1]:
    model = linear_model.ElasticNet(alpha = 1,
                                    l1_ratio = l1_ratio, 
                                    random_state = 123)
    model.fit(X_train.values, y_train)
    #make prediction
    pred = model.predict(X_test.values)
    pred = pred.clip(0, 20)

    # calculate score
    MSETest = mean_squared_error(y_test, pred)
    RMSETest = np.sqrt(MSETest)
    MSETrain = mean_squared_error(y_train, model.predict(X_train.values))
    RMSETrain = np.sqrt(MSETrain)
    
    st = datetime.datetime.fromtimestamp(time.time()).strftime('%m%d_%H.%M')
    joblib.dump(model, 'elasticNet_' + str(l1_ratio) + '_' + st + '.pkl')
    
    print('elasticNet_' + str(l1_ratio) + '_' + st + '.pkl')
    print('l1_ratio:', l1_ratio, 'Train RMSE:', RMSETrain, 'Test RMSE:', RMSETest)

#### Ridge model for the full training set

X_train_full = pd.read_pickle('../data/X_train_MinMaxScaler')
X_test_full = pd.read_pickle('../data/X_test_MinMaxScaler')
X_train_full = X_train_full.drop(['shop_id','item_id', 'item_category_id'], axis = 1)
X_test_full = X_test_full.drop(['shop_id','item_id', 'item_category_id'], axis = 1)

for alpha in [min(ridgeResult, key=ridgeResult.get)]:
    model = linear_model.Ridge(alpha = alpha, normalize = True, max_iter = 3000)
    model.fit(X_train_full.values, y_train_full)

st = datetime.datetime.fromtimestamp(time.time()).strftime('%m%d_%H.%M')
joblib.dump(model, 'ridge_full_' + st + '.pkl')
# ridge = joblib.load('ridge.pkl') 

submission(model, X_test_full)

#### Linear model for the full dataset
<br>The RMSE for this submission: 0.99202

X_train_full = pd.read_pickle('../data/X_train_RobustScaler')
X_test_full = pd.read_pickle('../data/X_test_RobustScaler')
X_train_full = X_train_full.drop(['shop_id','item_id', 'item_category_id'], axis = 1)
X_test_full = X_test_full.drop(['shop_id','item_id', 'item_category_id'], axis = 1)

model = linear_model.LinearRegression(n_jobs = 4)
model.fit(X_train_full, y_train_full)

st = datetime.datetime.fromtimestamp(time.time()).strftime('%m%d_%H.%M')
joblib.dump(model, 'lr_full_' + st + '.pkl')

submission(model, X_test_full)
