==> time-series-sales-prediction/Step4.3.Models-Neural-Networks.ipynb <==
# Sales Prediction for Time Series Data

## Part 4: Models (Neural Network)

import numpy as np
import pandas as pd 
import os
import time 
import datetime
import seaborn as sns
import matplotlib.pyplot as plt
import gc
from sklearn.metrics import mean_squared_error
from sklearn.externals import joblib
from keras.models import Sequential
from keras.layers import Dense, Activation
from keras.wrappers.scikit_learn import KerasRegressor
from keras.callbacks import ModelCheckpoint
from keras.callbacks import EarlyStopping
from keras.models import load_model
import h5py

%matplotlib inline 

def submission(model, X_test):
    '''
    make submission file
    arguments:    model = model name 
                  X_test= X_test name
    return: a file saved in directory with timestamp
    '''
    # model prediction
    pred = model.predict(X_test)
    print('mean before clipping: ', pred.mean())
    pred = pred.clip(0,20)
    print('mean after clipping: ', pred.mean())

    # create prediction dataframe
    ID = joblib.load('ID.pkl')
    predDF = pd.DataFrame() 
    predDF['ID'] = ID
    predDF['item_cnt_month'] = pred
    print(predDF.head())

    # write dataframe to csv
    ts = time.time()
    st = datetime.datetime.fromtimestamp(ts).strftime('%m%d_%H.%M')
    print('submission_' + st + '.csv')
    
    predDF.to_csv(header=True, index=False, path_or_buf = 'submission_' + st + '.csv')
    
    return None

DATA_FOLDER = '../data/'
X_train= pd.read_pickle(DATA_FOLDER+'X_train_lev_1_standardScaler')
y_train = joblib.load(DATA_FOLDER+'y_train_lev_1.pkl')
X_test = pd.read_pickle(DATA_FOLDER+'X_test_lev_1_standardScaler')
y_test = joblib.load(DATA_FOLDER+'y_val_lev_1.pkl')
X_train_full = pd.read_pickle(DATA_FOLDER+'X_train_standardScaler')
X_test_full = pd.read_pickle(DATA_FOLDER+'X_test_standardScaler')
y_train_full = joblib.load(DATA_FOLDER+'y_train_full.pkl')

Sample 20% of all items

sampleSize = 0.01 # 33 items
sample= list(X_train['item_id'].unique()) + list(X_test['item_id'].unique()) 
sample = list(set(sample))
np.random.seed(1234)
sample= list(np.random.choice(sample, size= int(len(sample)*sampleSize), replace=False, p=None))

Narrow down the training set.  

X_train = X_train.reset_index(drop = True)
X_test = X_test.reset_index(drop = True)

trainIndex = X_train.loc[(X_train['item_id'].isin(sample))].index
testIndex = X_test.loc[(X_test['item_id'].isin(sample))].index

X_train_s = X_train.loc[trainIndex]
X_test_s = X_test.loc[testIndex]
y_train_s = y_train[trainIndex]
y_test_s = y_test[testIndex]

Convert to arrays

X_train_s_m = X_train_s.as_matrix(columns= X_train_s.columns)
X_test_m = X_test.as_matrix(columns = X_test_s.columns)
X_train_m = X_train.as_matrix(columns= X_train.columns)
X_train_full_m = X_train_full.as_matrix(columns = X_train_full.columns)

print(X_train_s_m.shape)
print(X_test_m.shape)
print(y_train_s.size)
print(y_test.size)
print(X_train_full_m.shape)

Define base model

def base_model():
    # create model
    model = Sequential()
    model.add(Dense(64, input_dim=235, kernel_initializer='normal', activation='relu'))
    model.add(Dense(32, kernel_initializer='normal'))
    model.add(Dense(1, kernel_initializer='normal'))
    # Compile model
    model.compile(loss='mean_squared_error', optimizer='adam')
    return model

Define Model Checkpoint: saves the model weights after each epoch if the validation loss decreased

filepath= 'nnBestModel.hdf5'
checkpoint = ModelCheckpoint(filepath, 
                             monitor='val_loss', 
                             verbose=1, 
                             save_best_only=True, 
                             save_weights_only=False,
                             mode='auto')


Define Early Stopping: Stop training when a monitored quantity has stopped improving.

early_stop = EarlyStopping(monitor='val_loss', 
                           min_delta = 0.0001, #minimum change in the monitored quantity to qualify as an improvement,
                           patience= 5, #number of epochs with no improvement after which training will be stopped.
                           mode='auto') 

callbacks_list = [checkpoint, early_stop]
model = KerasRegressor(build_fn=base_model, verbose= True)
history = model.fit(X_train_m, y_train, validation_data=(X_test_m, y_test), epochs=50, batch_size=5, callbacks=callbacks_list)

Load previously trained model Weights (the best model from before)

model = load_model("nnBestModel.hdf5")
scores = model.evaluate(X_test_m, y_test)
print(scores)
MSE = mean_squared_error(y_test, model.predict(X_test_m))
RMSE = np.sqrt(MSE)
print('Test_sample MSE:', MSE)
print('Test_sample RMSE:', RMSE)

Submit Model (trained on subset of data)

submission(model, X_test_full) # 0.96402

#### Train on full set of data

model = KerasRegressor(build_fn=base_model, verbose= True)
model.fit(X_train_full_m, y_train_full, epochs = 1, batch_size=5)

submission(model, X_test_full) # 0.94596

model.model.save('nn_full_model.h5')
