==> time-series-sales-prediction/Step5.Ensembles.ipynb <==
# Sales Prediction for Time Series Data

This notebook gathers predictions from the three base models and train the stacked model.  For this project, the simple averaging of the model predictions did better than stacking.  

## Part 5: Ensembles

Predictions for date block 32 will be gathered from the models trained previously, which includes linear regression, light GBM, and neural networks.  

import numpy as np
import pandas as pd 
import os
import time 
import datetime
import seaborn as sns
import matplotlib.pyplot as plt
import gc
from sklearn import linear_model
from sklearn.metrics import mean_squared_error
from sklearn.externals import joblib
from keras.models import Sequential
from keras.layers import Dense, Activation
from keras.wrappers.scikit_learn import KerasRegressor
from keras.callbacks import ModelCheckpoint
from keras.callbacks import EarlyStopping
from keras.models import load_model
import h5py
import lightgbm as lgb

%matplotlib inline 

Load the base model files trained on date_block_num from 12 to 31

DATA_FOLDER = '../script/'
lrModel = joblib.load(DATA_FOLDER + 'lr_robust_scale_0201_15.05.pkl') 
lgbmModel = lgb.Booster(model_file='lgb_model_0131_17.30.txt') 
nnModel = load_model(DATA_FOLDER + "nnBestModel.hdf5")

Load the X_test and y_test for obtaining predictions.  

date_block_num = 32.  Used for meta model features

DATA_FOLDER = '../data/'
X_test_robustScaler = pd.read_pickle(DATA_FOLDER + 'X_test_lev_1_RobustScaler') # LR
X_test_standardScaler = pd.read_pickle(DATA_FOLDER + 'X_test_lev_1_standardScaler') # NNet
X_test_robustScalerTree = pd.read_pickle(DATA_FOLDER + 'X_test_lev_1_RobustScalerTree') #light GBM
y_test = joblib.load(DATA_FOLDER + 'y_val_lev_1.pkl')
X_test_robustScaler_lr = X_test_robustScaler.drop(['shop_id','item_id', 'item_category_id'], axis = 1)

date_block_num = 33.  Used for meta model validation

DATA_FOLDER = '../data/'
X_test_2_RobustScaler = pd.read_pickle(DATA_FOLDER + 'X_test_lev_2_RobustScaler') # LR
X_test_2_standardScaler = pd.read_pickle(DATA_FOLDER + 'X_test_lev_2_standardScaler') # NNet
X_test_2_robustScalerTree = pd.read_pickle(DATA_FOLDER + 'X_test_lev_2_RobustScalerTree') #light GBM
y_test_2 = joblib.load(DATA_FOLDER + 'y_val_lev_2.pkl')
X_test_2_robustScaler_lr = X_test_2_RobustScaler.drop(['shop_id','item_id', 'item_category_id'], axis = 1)

### Stacking

#### Gather predictions from base models. 
These are the meta features

lrPred = lrModel.predict(X_test_robustScaler_lr.values).clip(0,20)
lgbmPred = lgbmModel.predict(X_test_robustScalerTree.values).clip(0,20)
nnPred = nnModel.predict(X_test_standardScaler.values).clip(0,20)

lrPred_val = lrModel.predict(X_test_2_robustScaler_lr.values).clip(0,20)
lgbmPred_val = lgbmModel.predict(X_test_2_robustScalerTree.values).clip(0,20)
nnPred_val = nnModel.predict(X_test_2_standardScaler.values).clip(0,20)

preds = pd.DataFrame(columns = ['lr', 'lgbm', 'nnet'])
preds['lr'] = lrPred
preds['lgbm'] = lgbmPred
preds['nnet'] = nnPred
preds['mean'] = pd.DataFrame.mean(preds, axis = 1)

preds_val = pd.DataFrame(columns = ['lr', 'lgbm', 'nnet'])
preds_val['lr'] = lrPred_val
preds_val['lgbm'] = lgbmPred_val
preds_val['nnet'] = nnPred_val
preds_val['mean'] = pd.DataFrame.mean(preds_val, axis = 1)

X_train_meta = preds.iloc[:,0:4]
y_train_meta = y_test
X_test_meta = preds_val.iloc[:, 0:4]
y_test_meta = y_test_2

#### Baseline:  mean prediction from the base models

The base models were trained with data before date_block_num 32.  Below we calculate the MSE for block_num 32.  The meta model's training MSE should be better than these MSE.  

lrMSE = mean_squared_error(y_test, lrPred)
lgbmMSE = mean_squared_error(y_test, lgbmPred)
nnMSE = mean_squared_error(y_test, nnPred)
meanMSE = mean_squared_error(y_test, preds['mean'])
print('lr:', lrMSE, 'light gbm: ', lgbmMSE, 'neural netowrks:', nnMSE)
print('mean MSE of all models: ', meanMSE)

Date_block_num 33 is used for validation of the meta model.  Below we calculate the MSE using the base models for this block. The meta model's testing performance should be better than these numbers.   

lrMSE = mean_squared_error(y_test_2, lrPred_val)
lgbmMSE = mean_squared_error(y_test_2, lgbmPred_val)
nnMSE = mean_squared_error(y_test_2, nnPred_val)
meanMSE = mean_squared_error(y_test_2, preds_val['mean'])
print('lr:', lrMSE, 'light gbm: ', lgbmMSE, 'neural netowrks:', nnMSE)
print('mean MSE of all models: ',meanMSE)

#### Linear Regression

model = linear_model.LinearRegression(n_jobs = 4)
model.fit(X_train_meta, y_train_meta)
lr_meta_test_pred = model.predict(X_test_meta.values)
lr_meta_train_pred = model.predict(X_train_meta.values)
lr_meta_test_MSE = mean_squared_error(y_test_meta, lr_meta_test_pred.clip(0, 20))
lr_meta_train_MSE = mean_squared_error(y_train_meta, lr_meta_train_pred.clip(0, 20))
print('Linear Regression Meta Model Train (block 32) MSE: ',lr_meta_train_MSE)
print('Linear Regression Meta Model Test (block 33) MSE: ',lr_meta_test_MSE)

st = datetime.datetime.fromtimestamp(time.time()).strftime('%m%d_%H.%M')
joblib.dump(model, 'lr_meta_' + st + '.pkl')

#### Light GBM

def lgb_model(params):
    lgb_model = lgb.train(params, train_data, valid_sets=[train_data, test_data], verbose_eval=25)
    return lgb_model

params = {
    'application':'regression',
    'learning_rate':0.001,
    'early_stopping_round':10,
    'metric':'l2_root', #RMSE
    'nthread':-1, 
    'train_metric': True,
    'num_boost_round': 1000,
    'max_depth:': 2
}

train_data = lgb.Dataset(X_train_meta, label= y_train_meta)
test_data = lgb.Dataset(X_test_meta, label = y_test_meta)

model = lgb_model(params) 

print(model.best_iteration)

lgbm_meta_test_pred = model.predict(X_test_meta.values)
lgbm_meta_train_pred = model.predict(X_train_meta.values)
lgbm_meta_test_MSE = mean_squared_error(y_test_meta, lgbm_meta_test_pred.clip(0, 20))
lgbm_meta_train_MSE = mean_squared_error(y_train_meta, lgbm_meta_train_pred.clip(0, 20))
print('Light GBM Meta Model Train (block 32) MSE: ',lgbm_meta_train_MSE)
print('Light GBM Meta Model Test (block 33) MSE: ',lgbm_meta_test_MSE)

st = datetime.datetime.fromtimestamp(time.time()).strftime('%m%d_%H.%M')
model.save_model('lgb_meta_' + st + '.txt', num_iteration= model.best_iteration)
print('lgb_meta_' + st + '.txt')

### Final Submission

Load models that are trained on the full dataset.

DATA_FOLDER = '../script/'
lrModel = joblib.load(DATA_FOLDER + 'lr_full_0201_21.06.pkl') 
lgbmModel = lgb.Booster(model_file ='lgb_model_full_0131_17.52.txt') 
nnModel = load_model(DATA_FOLDER + 'nn_full_model.h5')
lgb_meta = lgb.Booster(model_file = DATA_FOLDER + 'lgb_meta_0201_21.13.txt')
lr_meta = joblib.load(DATA_FOLDER + 'lr_meta_0201_21.09.pkl') 

DATA_FOLDER = '../data/'
X_test_robustScaler = pd.read_pickle(DATA_FOLDER + 'X_test_RobustScaler') # LR
X_test_standardScaler = pd.read_pickle(DATA_FOLDER + 'X_test_standardScaler') # NNet
X_test_robustScalerTree = pd.read_pickle(DATA_FOLDER + 'X_test_RobustScalerTree') #light GBM
X_test_robustScaler_lr = X_test_robustScaler.drop(['shop_id','item_id', 'item_category_id'], axis = 1)

Get predictions for the base learners

lrFinalPred = lrModel.predict(X_test_robustScaler_lr).clip(0,20)
lgbmFinalPred = lgbmModel.predict(X_test_robustScalerTree).clip(0,20)
nnFinalPred = nnModel.predict(X_test_standardScaler).clip(0,20)

preds_test = pd.DataFrame(columns = ['lr', 'lgbm', 'nnet'])
preds_test['lr'] = lrFinalPred
preds_test['lgbm'] = lgbmFinalPred 
preds_test['nnet'] = nnFinalPred
preds_test['mean'] = pd.DataFrame.mean(preds_test, axis = 1)

preds_test.shape

Get predictions for the meta learners

lrMetaPred = lr_meta.predict(preds_test).clip(0,20)
lgbMetaPred = lgb_meta.predict(preds_test).clip(0,20)

def submit_final_pred(pred):
    ID = joblib.load('ID.pkl')
    predDF = pd.DataFrame() 
    predDF['ID'] = ID
    predDF['item_cnt_month'] = pred
    ts = time.time()
    st = datetime.datetime.fromtimestamp(ts).strftime('%m%d_%H.%M')
    print('submission_' + st + '.csv')

    predDF.to_csv(header=True, index=False, path_or_buf = 'submission_' + st + '.csv')
    
    return None

submit_final_pred(lrMetaPred) # 1.00192

submit_final_pred(lgbMetaPred) # 0.97329

submit_final_pred(preds_test['mean']) # 0.92366
